BackEnd/requirements.txt
asgiref==3.3.1
Django==3.0.5
django-cors-headers==3.5.0
djangorestframework==3.12.2
djongo==1.3.3
dnspython==2.0.0
pymongo==3.11.2
pytz==2020.4
sqlparse==0.2.4
